update to 1.4.7

+++ src/FisherScoring.h
@@ -13,7 +13,7 @@ extern "C" {
 }
 #endif
 using namespace std;
-void GetSAndH(float* data, float* data2, double* beta, int length, int* labels, int rank, double* S, double* H, int v1, int v2, int n);
-double* GetInverseMat(double* mat, int rank);
-double DoIteration(float* data, float* data2, int length, int* labels, double epsilon, int rank, int v1, int v2);
-double DoIteration2(float* dataHead, int offset1, float* dataHead2, int offset2, int length, int* labels, double epsilon, int rank);
+void GetSAndH(float* data, float* data2, double* beta, int length, int* labels, int rank, float* S, float* H, int v1, int v2, int n);
+float* GetInverseMat(float* mat, int rank);
+float DoIteration(float* data, float* data2, int length, int* labels, float epsilon, int rank, int v1, int v2);
+float DoIteration2(float* dataHead, int offset1, float* dataHead2, int offset2, int length, int* labels, float epsilon, int rank);
